requirements.txt
# ============================================================================
# GenCP Requirements
# ============================================================================
# 
# Installation Instructions:
# 1. Install PyTorch via conda (recommended for CUDA support):
#    conda install pytorch=2.2.0 torchvision=0.17.0 torchaudio=2.2.0 pytorch-cuda=12.1 -c pytorch -c nvidia
#
# 2. Install other dependencies:
#    pip install -r requirements.txt
#
# 3. Install the package:
#    pip install -e .
#
# ============================================================================

# ----------------------------------------------------------------------------
# Core Deep Learning Framework
# ----------------------------------------------------------------------------
# Note: PyTorch should be installed via conda for proper CUDA support
# If installing via pip, uncomment the following lines:
# torch>=2.0.0
# torchvision>=0.15.0
# torchaudio>=2.0.0

# ----------------------------------------------------------------------------
# Numerical Computing and Scientific Libraries
# ----------------------------------------------------------------------------
numpy==1.26.3
scipy==1.14.1
sympy==1.12
mpmath==1.3.0

# ----------------------------------------------------------------------------
# Data Loading and Processing
# ----------------------------------------------------------------------------
h5py>=3.8.0  # Required for dataset loading (ntcouple, turek_hron, double_cylinder)
netCDF4==1.7.1.post2
pandas==2.2.2

# ----------------------------------------------------------------------------
# Configuration Management
# ----------------------------------------------------------------------------
PyYAML==6.0.2
yacs==0.1.8

# ----------------------------------------------------------------------------
# Flow Matching and ODE Solvers
# ----------------------------------------------------------------------------
torchdiffeq==0.2.5  # ODE solver for inference
POT==0.9.5  # Python Optimal Transport (required by torchcfm/optimal_transport.py)
# torchdyn>=1.0.0  # Optional: Only used in torchcfm/utils.py for demo/visualization

# Note: torchcfm is included in this repository (GenCP/torchcfm/) 
# and does not need to be installed separately

# ----------------------------------------------------------------------------
# Model Architectures and Neural Network Components
# ----------------------------------------------------------------------------
einops==0.8.0  # Tensor operations (rearrange, repeat)
einops-exts==0.0.4
timm==1.0.9  # Vision Transformer components (PatchEmbed, Attention, Mlp)
rotary-embedding-torch==0.6.5  # Rotary position embeddings

# ----------------------------------------------------------------------------
# Training and Distributed Computing
# ----------------------------------------------------------------------------
accelerate==0.33.0  # Multi-GPU training support
tensorboard==2.17.1  # Optional: Training visualization
tensorboard-data-server==0.7.2
ema-pytorch==0.6.0  # Exponential moving average for model weights

# ----------------------------------------------------------------------------
# Visualization
# ----------------------------------------------------------------------------
matplotlib==3.9.2
seaborn>=0.12.0  # Used in toy_our.py and toy_m2pde.py
pillow==10.2.0
plotly==5.24.0

# ----------------------------------------------------------------------------
# Utilities
# ----------------------------------------------------------------------------
tqdm==4.66.5
termcolor==2.4.0
tenacity==9.0.0

# ----------------------------------------------------------------------------
# Wavelet Transforms
# ----------------------------------------------------------------------------
pytorch-wavelets==1.3.0
PyWavelets==1.9.0
ptwt==1.0.0

# ----------------------------------------------------------------------------
# Optional Dependencies (uncomment if needed)
# ----------------------------------------------------------------------------
# xformers>=0.0.20  # Optional: Memory-efficient attention for latte model
# jax>=0.4.0  # Optional: Only used in CNO_libs/local_torch_utils/res_utils.py
# jaxlib>=0.4.0

# ----------------------------------------------------------------------------
# HTTP and Network Libraries
# ----------------------------------------------------------------------------
requests==2.32.3
aiohttp==3.10.5
aiohappyeyeballs==2.4.0
aiosignal==1.3.1
urllib3==2.2.2
yarl==1.9.4

# ----------------------------------------------------------------------------
# HuggingFace Ecosystem
# ----------------------------------------------------------------------------
huggingface-hub==0.24.6
safetensors==0.4.4

# ----------------------------------------------------------------------------
# Machine Learning Utilities
# ----------------------------------------------------------------------------
scikit-learn==1.5.1
dgl==2.1.0  # Deep Graph Library

# ----------------------------------------------------------------------------
# Development and Testing (Optional)
# ----------------------------------------------------------------------------
pytest==8.3.3
setuptools==72.1.0
wheel==0.43.0

# ----------------------------------------------------------------------------
# Additional Standard Libraries and Dependencies
# ----------------------------------------------------------------------------
absl-py==2.1.0
attrs==24.2.0
certifi==2024.7.4
cftime==1.6.4
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
fastjsonschema==2.20.0
filelock==3.13.1
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.2.0
grpcio==1.66.0
idna==3.8
iniconfig==2.0.0
Jinja2==3.1.3
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
Markdown==3.7
MarkupSafe==2.1.5
multidict==6.0.5
nbformat==5.10.4
networkx==3.2.1
pluggy==1.5.0
protobuf==5.27.3
pyparsing==3.1.3
pytz==2024.1
referencing==0.35.1
rpds-py==0.20.0
threadpoolctl==3.5.0
torchdata==0.8.0
typing_extensions==4.9.0
tzdata==2024.1
Werkzeug==3.0.4

# Windows-specific (commented out for Linux)
# pywin32==305.1
